dom/base/test/test_nondomexception.html

<!-- intentionally omiting doctype because this test requires document.all -->
<html>
<head>
  <title>Test for non-DOM module exceptions</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="application/javascript;version=1.7">
"use strict";

try {
  document.all();
} catch (e) {
  is(typeof e, "object");
  is(e.filename, location);
  is(e.lineNumber, 18);
}

</script>
</pre>
</body>
</html>
